@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <LangVersion>7.1</LangVersion>
     <!--
-      This is required for compiling razor views in a console app.
-      It is supposed to be fixed in 2.1, so we can remove it then.
-      https://github.com/aspnet/Razor/issues/1212
-    -->
+  This is required for compiling razor views in a console app.
+  It is supposed to be fixed in 2.1, so we can remove it then.
+  https://github.com/aspnet/Razor/issues/1212
+-->
     <PreserveCompilationContext>true</PreserveCompilationContext>
   </PropertyGroup>
   <ItemGroup>
@@ -15,7 +15,6 @@
     <ProjectReference Include="..\..\src\Statik\Statik.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Markdig" Version="0.15.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.1.1" />
+    <PackageReference Include="Markdig" Version="0.18.0" />
   </ItemGroup>
 </Project>
